requirements.txt
# Internal Knowledge Assistant — pinned dependencies

# UI
streamlit==1.56.0
streamlit-extras==1.5.0

# LangChain stack
langchain==1.2.15
langchain-core==1.3.0
langchain-classic==1.0.4
langchain-openai==1.2.0
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-text-splitters==1.1.2

# LLM / embeddings client
openai==2.32.0
tiktoken==0.12.0

# Vector store
chromadb==1.5.8

# Document loaders
pymupdf==1.27.2.2
pymupdf4llm==1.27.2.2

# Retrieval
rank-bm25==0.2.2

# Config
pydantic-settings==2.14.0
python-dotenv==1.2.2
